+++ libraries/I2C_LCD/I2C_LCD_special_chars.h
@@ -1,0 +1,31 @@
+#pragma once
+//
+//    FILE: I2C_LCD_special_chars.h
+//  AUTHOR: Rob.Tillaart@gmail.com
+// VERSION: see library.properties
+//    DATE: 2023-12-16
+// PUPROSE: Arduino library for I2C_LCD
+//     URL: https://github.com/RobTillaart/I2C_LCD
+
+
+//  SPECIAL CHARS
+
+const char LCD_ALPHA    = 0xE0;
+const char LCD_BETA     = 0xE2;
+const char LCD_EPSILON  = 0xE3;
+const char LCD_MU       = 0xE4;
+const char LCD_RHO      = 0xE5;
+const char LCD_THETA    = 0xF2;
+const char LCD_INFINITY = 0xF3;
+const char LCD_OHM      = 0xF4;
+const char LCD_SIGMA    = 0xF6;
+const char LCD_PI       = 0xF7;
+const char LCD_XAVG     = 0xF8;
+
+const char LCD_DEGREE   = 0xDF;
+const char LCD_DIVIDE   = 0xFD;
+
+
+
+//  -- END OF FILE --
+